--- kmp/kmp-htmlparser/build.gradle.kts
@@ -1,3 +1,0 @@
-plugins {
-    id("kmp-module-plugin")
-}
