+++ kmp/kmp-search-data/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("kmp-data-plugin")
+}
