+++ kmp/kmp-search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/Post.kt
@@ -1,0 +1,7 @@
+package com.gchristov.thecodinglove.kmpsearchdata
+
+data class Post(
+    val title: String,
+    val url: String,
+    val imageUrl: String,
+)
