+++ kmp/kmp-search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/parse/PostParser.kt
@@ -1,0 +1,8 @@
+package com.gchristov.thecodinglove.kmpsearchdata.parse
+
+import com.gchristov.thecodinglove.kmpsearchdata.Post
+
+interface PostParser {
+    fun parse(): String
+    fun parse(content: String): List<Post>
+}
