+++ modules.gradle.kts
@@ -8,8 +8,8 @@ val projects = listOf(
     "kmp-common-kotlin",
     "kmp-common-network",
     "kmp-common-test",
-    "kmp-htmlparser",
     "kmp-module-b",
+    "kmp-search-data",
     "appApiJs",
     "appHtmlJs",
     "gradle-plugins",
